Put global bool config in own database module
