+++ rust/src/database.rs
@@ -1,21 +1,20 @@
 //! Module for interacting with redb database, to store high level state, and non sensitive data.
 //! That will be available across the app, and will be persisted across app launches.
 
+pub mod global_bool;
 pub mod wallet;
 
 use std::{path::PathBuf, sync::Arc};
 
-use bdk_wallet::bitcoin::Network;
-use eyre::{Context, Result};
+use global_bool::GlobalBoolTable;
+use wallet::WalletTable;
+
+use eyre::Context;
 use log::{debug, error, info};
 use once_cell::sync::OnceCell;
 use redb::TableDefinition;
-use wallet::WalletTable;
 
-use crate::{
-    update::{Update, Updater},
-    view_model::wallet::WalletId,
-};
+use crate::view_model::wallet::WalletId;
 
 pub static DATABASE: OnceCell<Database> = OnceCell::new();
 
@@ -24,30 +23,10 @@ pub const GLOBAL_BOOL_CONFIG: TableDefinition<&'static str, bool> =
 
 pub const WALLETS: TableDefinition<&'static str, Vec<WalletId>> = TableDefinition::new("wallets");
 
-#[derive(Debug, Clone, Copy, strum::IntoStaticStr, uniffi::Enum)]
-pub enum GlobalBoolConfigKey {
-    CompletedOnboarding,
-}
-
-#[derive(Debug, Clone, Copy, strum::IntoStaticStr, uniffi::Enum)]
-pub enum WalletKey {
-    Bitcoin,
-    Testnet,
-}
-
-impl From<Network> for WalletKey {
-    fn from(network: Network) -> Self {
-        match network {
-            Network::Bitcoin => WalletKey::Bitcoin,
-            Network::Testnet => WalletKey::Testnet,
-            other => panic!("unsupported network: {other:?}"),
-        }
-    }
-}
-
 #[derive(Debug, Clone, uniffi::Object)]
 pub struct Database {
     pub db: Arc<redb::Database>,
+    pub global_bool: GlobalBoolTable,
     pub wallets: WalletTable,
 }
 
@@ -91,61 +70,6 @@ impl Database {
     pub fn new() -> Self {
         Self::global().clone()
     }
-
-    pub fn get_bool_config(&self, key: GlobalBoolConfigKey) -> Result<bool, Error> {
-        let read_txn = self
-            .db
-            .begin_read()
-            .map_err(|error| Error::DatabaseAccessError(error.to_string()))?;
-
-        let table = read_txn
-            .open_table(GLOBAL_BOOL_CONFIG)
-            .map_err(|error| Error::TableAccessError(error.to_string()))?;
-
-        let key: &'static str = key.into();
-        let value = table
-            .get(key)
-            .map_err(|error| Error::ConfigReadError(error.to_string()))?
-            .map(|value| value.value())
-            .unwrap_or(false);
-
-        Ok(value)
-    }
-
-    pub fn set_bool_config(&self, key: GlobalBoolConfigKey, value: bool) -> Result<(), Error> {
-        let write_txn = self
-            .db
-            .begin_write()
-            .map_err(|error| Error::DatabaseAccessError(error.to_string()))?;
-
-        {
-            let mut table = write_txn
-                .open_table(GLOBAL_BOOL_CONFIG)
-                .map_err(|error| Error::TableAccessError(error.to_string()))?;
-
-            let key: &'static str = key.into();
-            table
-                .insert(key, value)
-                .map_err(|error| Error::ConfigSaveError(error.to_string()))?;
-        }
-
-        write_txn
-            .commit()
-            .map_err(|error| Error::DatabaseAccessError(error.to_string()))?;
-
-        Updater::send_update(Update::DatabaseUpdate);
-
-        Ok(())
-    }
-
-    pub fn toggle_bool_config(&self, key: GlobalBoolConfigKey) -> Result<(), Error> {
-        let value = self.get_bool_config(key)?;
-
-        let new_value = !value;
-        self.set_bool_config(key, new_value)?;
-
-        Ok(())
-    }
 }
 
 impl Database {
@@ -156,8 +80,13 @@ impl Database {
 
             let db = Arc::new(db);
             let wallets = WalletTable::new(db.clone());
+            let global_bool = GlobalBoolTable::new(db.clone());
 
-            Database { db, wallets }
+            Database {
+                db,
+                wallets,
+                global_bool,
+            }
         })
     }
 }
